Attempt to fix persistent regression (#165)
I have regrettably been unable to reproduce the bug on our side with the
broken version, but I have updated the build files to change it to the
version with the fix.

The stack build is mostly untested because stack crashes halfway through
any build with an encoding error regardless of what locale overrides I
give it.